pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "event manager django"
authors = [
    {name = "Julia Alexeeva",email = "yulia.alekseeva12@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-filter (>=25.1,<26.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
mypy = "^1.15.0"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
